MZSHUIYIN/requirements.txt
allure_python_commons==2.13.2
selenium==4.11.2
pyautogui==0.9.54
pypiwin32
pytest==7.4.0
pytest-html==3.2.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
undetected_chromedriver==3.5.3
pytest-rerunfailures
pytest-xlsx==0.5.2
pytest-xdist==3.3.1
webdriver_manager==4.0.0
numpy==1.25.2
opencv-python==4.8.0.76
pynput==1.7.6